requirements.txt
streamlit
pandas
scikit-learn
matplotlib
pyzbar
Pillow
opencv-python-headless
